♻️ refactor(typefile, AuthContext): 타입파일 이름 변경, 컨텍스트 useEffect기반으로 변경
