+++ src/context/AuthContext.tsx
@@ -1,10 +1,9 @@
 "use client";
 
-import { createContext, useContext, useState, ReactNode, useMemo } from "react";
-
+import { createContext, useContext, useState, ReactNode, useMemo, useEffect } from "react";
 import getUsersMe from "@/apis/user/getUsersMe";
-import { IUser } from "@/types/user";
 import getProfilesCode, { RequestProfileCode } from "@/apis/profile/getProfilesCode";
+import { IUser } from "@/types/user.type";
 
 interface AuthContextType {
   user: IUser | null;
@@ -12,7 +11,6 @@ interface AuthContextType {
   getUser: () => void;
 }
 
-// const AuthContext = createContext<AuthContextType | undefined>(undefined);
 const AuthContext = createContext<AuthContextType>({
   user: null,
   userProfile: null,
@@ -20,39 +18,38 @@ const AuthContext = createContext<AuthContextType>({
 });
 
 function AuthProvider({ children }: { children: ReactNode }) {
-  // const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [user, setUser] = useState<IUser | null>(null);
   const [userProfile, setUserProfile] = useState<RequestProfileCode | null>(null);
 
-  const getUserProfile = async (newUser: IUser | null) => {
-    if (newUser?.profile) {
-      const newUserProfile = await getProfilesCode(newUser.profile.code);
-      console.log("$$ newUserProfile", newUserProfile);
-      setUserProfile(() => {
-        if (newUserProfile) return newUserProfile;
-        return null;
-      });
-    } else {
-      setUserProfile(() => null);
-    }
-  };
-
-  const values = useMemo(() => {
-    const getUser = async () => {
-      const newUser = await getUsersMe();
-      setUser(() => {
-        if (newUser) {
-          (async () => {
-            await getUserProfile(newUser);
-          })();
-          return newUser;
-        }
-        return null;
-      });
+  useEffect(() => {
+    const getUserProfile = async (newUser: IUser | null) => {
+      if (newUser?.profile) {
+        const newUserProfile = await getProfilesCode(newUser.profile.code);
+        setUserProfile(newUserProfile);
+      } else {
+        setUserProfile(null);
+      }
     };
 
-    return { user, userProfile, getUser };
-  }, [user, userProfile]);
+    const getUser = async () => {
+      const newUser = await getUsersMe();
+      setUser(newUser);
+      if (newUser) {
+        await getUserProfile(newUser);
+      }
+    };
+
+    getUser();
+  }, []); // 빈 배열을 두어 컴포넌트가 마운트될 때 한 번만 실행되도록 한다
+
+  const values = useMemo(
+    () => ({
+      user,
+      userProfile,
+      getUser: () => {},
+    }),
+    [user, userProfile],
+  );
 
   return <AuthContext.Provider value={values}>{children}</AuthContext.Provider>;
 }
